Updated TCC calculations
- Fixed all various issues with both TCC calculations (numba version and C version).
- Also changed the .cpp file such that users don't have to manually comment out the declspec block.
- Some cleanup of small linting issues (shouldn't impact runtime performance)
- removed the heisenbug of overflow/underflow/invalid values in arrays. Didn't impact the outputs of the program, but it is nice to have the warning removed nonetheless.
